// src/models/components/createtransferresponsetype.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

import * as z from "zod";
import { ClosedEnum } from "../../types/enums.js";

export const CreateTransferResponseType = {
  CreatedTransfer: "CreatedTransfer",
  AsyncTransfer: "AsyncTransfer",
  Transfer: "Transfer",
} as const;
export type CreateTransferResponseType = ClosedEnum<
  typeof CreateTransferResponseType
>;

/** @internal */
export const CreateTransferResponseType$inboundSchema: z.ZodNativeEnum<
  typeof CreateTransferResponseType
> = z.nativeEnum(CreateTransferResponseType);

/** @internal */
export const CreateTransferResponseType$outboundSchema: z.ZodNativeEnum<
  typeof CreateTransferResponseType
> = CreateTransferResponseType$inboundSchema;

/**
 * @internal
 * @deprecated This namespace will be removed in future versions. Use schemas and types that are exported directly from this module.
 */
export namespace CreateTransferResponseType$ {
  /** @deprecated use `CreateTransferResponseType$inboundSchema` instead. */
  export const inboundSchema = CreateTransferResponseType$inboundSchema;
  /** @deprecated use `CreateTransferResponseType$outboundSchema` instead. */
  export const outboundSchema = CreateTransferResponseType$outboundSchema;
}